+++ changelog.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
+### Added
+
+- Private admin command `/excluded` for listing excluded tournaments eligible for the next poll.
+
 ## [0.1.0] - 2026-07-17
 
 ### Added